FT Cat State Prep (#378)
## Description

This PR adds some minor functionality to test repeat-until-success
preparation of cat states.

Furthermore, this PR introduces the `noise` module, which provides
functionality for applying noise models to stim circuits and defining
new noise models. This is in a rudimentary stage for now and is only
applied to the cat state simulations. In the future, I would like to
extend this to more noise models and remove redundant code in the
simulation of state preparation circuits.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: Tom Peham <tom.peham@tum.de>
Co-authored-by: Erik Weilandt <erik.weilandt@tum.de>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>