Decouple Fault Set functionality from state preparation. (#443)
<!--- This file has been generated from an external template. Please do
not modify it directly. -->
<!--- Changes should be contributed to
https://github.com/munich-quantum-toolkit/templates. -->
## Description

In our works on state preparation circuit synthesis we use the concept
of the fault set of a circuit, which is just the collection of all
circuit faults propagated to the output. For the circuit sizes we are
interested in it is nice to work with this explicit collection as it
makes it easy to construct verification circuits.

In the current implementation, all this functionality is tightly coupled
to the `StatePrepCircuit` class. Since we want to analyze circuits
beyond state preparation circuits, it makes sense to decouple this a
little bit. Therefore, this PR introduces a `PureFaultSet` class which
represents collections of faults of a single type (X or Z). Many
functions surrounding verification circuit synthesis now take fault sets
as inputs instead of circuits. The `StatePrepCircuit` class still exists
to cache computed fault sets for state preparation circuits and to
guarantee the existence of verification circuits.

This PR also introduces the `CNOTCircuit` class as an intermediate
representation during the synthesis to simplify working with these kinds
of circuits. The idea is to have CNOT circuits with arbitrary fixed
inputs to represent any CSS encoding isometry.

Furthermore, this PR refactors some of the state preparation circuit
synthesis code.

List of changes:

- New class `PureFaultSet` for representing collections of X or Z faults
- New class `CNOTCircuit` 
- Renamed `StatePrepCircuit` class to `FaultyStatePrepCircuit`. This
makes it clearer that this is an object that combines information about
a circuit and its faults.
- Refactored state preparation circuit synthesis code to use the new
features.

Fixes #438

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [ ] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: Tom Peham <tom.peham@tum.de>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>